+++ ADPraktikum/src/Aufgabe_12/aufgabe12/src/Aufgabe_12/Gauss.java
@@ -1,4 +1,4 @@
-package Aufgabe_7;
+package Aufgabe_12;
 
 public class Gauss {
