+++ ADPraktikum/src/Aufgabe_12/aufgabe12/src/Aufgabe_12/Markovseq.java
