+++ ADPraktikum/src/Aufgabe_12/aufgabe12/src/Aufgabe_12/Rational.java
@@ -1,4 +1,4 @@
-package Aufgabe_7;
+package Aufgabe_12;
 
 public class Rational {
 	public long zaehler;
