+++ ADPraktikum/src/Aufgabe_12/aufgabe12/src/Aufgabe_12/StdRandom.java
@@ -1,4 +1,4 @@
-package Aufgabe_7;
+package Aufgabe_12;
 /******************************************************************************
  *  Compilation:  javac StdRandom.java
  *  Execution:    java StdRandom
